SLICE working through python with hard-coded parameters
