+++ lib/gamtools/slice.py
@@ -1,0 +1,5 @@
+from . import slice_wrapper
+
+def run_slice_from_args(args):
+    print('Running slice...')
+    slice_wrapper.slice()
