+++ lib/gamtools/slice_wrapper.pyx
@@ -1,0 +1,5 @@
+cdef extern from "slice.cpp":
+    int run_slice()
+
+def slice():
+    return run_slice()
